#5 + prepare global abstract network wrapper (with common methods for HTTP and RT protocols) + unit-test improvements
