+++ test/network/query.test.js
@@ -2,17 +2,18 @@ var
   should = require('should'),
   sinon = require('sinon'),
   rewire = require('rewire'),
+  AbstractWrapper = rewire('../../src/networkWrapper/protocols/abstract/common'),
   RTWrapper = rewire('../../src/networkWrapper/protocols/abstract/realtime');
 
 describe('Network query management', function () {
   describe('#emitRequest', function () {
     var
-      emitRequest = RTWrapper.__get__('emitRequest'),
+      emitRequest = AbstractWrapper.__get__('emitRequest'),
       sendSpy,
       network;
 
     beforeEach(function () {
-      network = new RTWrapper('somewhere');
+      network = new AbstractWrapper('somewhere');
       network.send = function(request) {
         network.emit(request.requestId, request.response);
       };
@@ -94,7 +95,7 @@ describe('Network query management', function () {
 
     before(function () {
       emitRequestStub = sinon.stub();
-      emitRequestRevert = RTWrapper.__set__('emitRequest', emitRequestStub);
+      emitRequestRevert = AbstractWrapper.__set__('emitRequest', emitRequestStub);
     });
 
     after(function () {
@@ -102,7 +103,7 @@ describe('Network query management', function () {
     });
 
     beforeEach(function () {
-      network = new RTWrapper('somewhere');
+      network = new AbstractWrapper('somewhere');
       network.state = 'connected';
       emitRequestStub.reset();
     });
@@ -261,6 +262,18 @@ describe('Network query management', function () {
       response = {result: {channel: 'foobar', roomId: 'barfoo'}};
     });
 
+    it('should throw an error if the protocol does not support realtime', function() {
+      var cb = sinon.stub();
+
+      network = new AbstractWrapper('somewhere');
+
+      network.subscribe({}, {}, sinon.stub(), cb);
+      should(cb).be.calledOnce();
+      should(cb.firstCall.args[0]).be.an.instanceOf(Error);
+      should(cb.firstCall.args[0].message).be.eql('Not Implemented');
+      should(network.query).not.be.called();
+    });
+
     it('should throw an error if not connected', function() {
       var cb = sinon.stub();
 
@@ -370,6 +383,18 @@ describe('Network query management', function () {
       network.query = sinon.stub().callsFake(queryStub);
       error = null;
       response = {result: {roomId: 'foobar'}};
+    });
+
+    it('should throw an error if the protocol does not support realtime', function() {
+      var cb = sinon.stub();
+
+      network = new AbstractWrapper('somewhere');
+
+      network.unsubscribe({foo: 'bar'}, {bar: 'foo'}, 'channel', cb);
+      should(cb).be.calledOnce();
+      should(cb.firstCall.args[0]).be.an.instanceOf(Error);
+      should(cb.firstCall.args[0].message).be.eql('Not Implemented');
+      should(network.query).not.be.called();
     });
 
     it('should call query method with good arguments', function() {
